requirements.txt
ollama
openai
httpx
google-genai
omegaconf
sounddevice
faster-whisper
silero-vad
edge-tts
numpy
pydantic
python-dotenv
rich
piper-tts
